refactor: separate migration from cli
